src/proexe_assignment/error_handling.py
from rest_framework.views import exception_handler
from django.db.utils import IntegrityError
from rest_framework.response import Response
from rest_framework import status


def api_exception_handler(exc, context):
    response = exception_handler(exc, context)

    response.data = {"errors":response.data}

    return response
